#13663: Relocate math.hpp
