+++ ttnn/cpp/ttnn/operations/data_movement/pad/device/pad_program_factory.cpp
@@ -4,7 +4,7 @@
 
 #include "ttnn/tensor/host_buffer/functions.hpp"
 #include "tt_metal/common/work_split.hpp"
-#include "ttnn/deprecated/tt_dnn/op_library/math.hpp"
+#include "ttnn/operations/moreh/math.hpp"
 #include "tt_metal/common/constants.hpp"
 #include "tt_metal/detail/util.hpp"
 #include "tt_metal/host_api.hpp"
