+++ ttnn/cpp/ttnn/operations/pool/downsample/device/downsample_op.cpp
@@ -7,7 +7,7 @@
 
 #include <math.h>
 
-#include "ttnn/deprecated/tt_dnn/op_library/math.hpp"
+#include "ttnn/operations/moreh/math.hpp"
 #include "tt_metal/common/constants.hpp"
 
 using namespace tt::constants;
