+++ ttnn/cpp/ttnn/operations/pool/downsample/device/downsample_program_factory.cpp
@@ -7,7 +7,7 @@
 #include <math.h>
 
 #include "ttnn/common/constants.hpp"
-#include "ttnn/deprecated/tt_dnn/op_library/math.hpp"
+#include "ttnn/operations/moreh/math.hpp"
 #include "tt_metal/common/work_split.hpp"
 #include "ttnn/operations/data_movement/untilize/untilize.hpp"
 #include "tt_metal/common/constants.hpp"
